+++ src/commands/browser/stagehand/initOverride.ts
@@ -267,7 +267,9 @@ const oldInit = Stagehand.prototype.init;
 export const patchStagehand = once(async () => {
   const tempStagehand = new Stagehand({
     env: 'LOCAL',
-    apiKey: 'test',
+    modelClientOptions:{
+      apiKey: 'test',
+    },
     verbose: 0,
     localBrowserLaunchOptions: {
       headless: true,
@@ -275,6 +277,7 @@ export const patchStagehand = once(async () => {
     logger: () => {},
   });
   await oldInit.call(tempStagehand);
+  type StagehandPageType = Parameters<typeof tempStagehand['setActivePage']>[0]
   const StagehandPage = tempStagehand['stagehandPage']!.constructor;
   const StagehandContext = tempStagehand['stagehandContext']!.constructor;
   tempStagehand.context.addInitScript;
@@ -339,6 +342,10 @@ export function overrideStagehandInit() {
         const defaultPage =
           pages.length > 0 ? pages[pages.length - 1] : await browserResult.context.newPage();
 
+          if(!this['llmClient']){
+            throw new Error('llmClient not found. This is usually because the provider or model name is not recognized by this version of stagehand.')
+          }
+        // constructor(page: Page$1, stagehand: Stagehand, context: StagehandContext, llmClient: LLMClient, userProvidedInstructions?: string, api?: StagehandAPI, waitForCaptchaSolves?: boolean);
         this['stagehandPage'] = await new StagehandPage(
           defaultPage,
           this,